refactor(redispositioner): rename package and update module paths
