+++ mysql/canal/positioner/redis/options.go
@@ -1,4 +1,4 @@
-package redispositioner
+package redis
 
 import (
 	"strings"
