--- mysql/canal/server.go
@@ -1,1 +1,0 @@
-package canal
