+++ versions.yaml
@@ -89,7 +89,8 @@ module-sets:
 
       # MySQL
       - github.com/go-fries/fries/mysql/canal/v3
-      - github.com/go-fries/fries/mysql/canal/redispositioner/v3
+      - github.com/go-fries/fries/mysql/canal/positioner/redis/v3
+      - github.com/go-fries/fries/mysql/canal/server/v3
 
       # Otel
       - github.com/go-fries/fries/otel/otlp/v3
